requirements.txt
# Core dependencies
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
groq>=0.4.1

# RAG & Vector Store
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# Document Processing
pypdf>=3.17.0

# HTTP & API
requests>=2.31.0

# Environment Management
python-dotenv>=1.0.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
